Early version of selection task is done
As of now, one can call the RandomWindowOn() method multiple times to
turn on a window in the 2x3 flat layout. Most importantly, the user can
only select these windows (the ones that are turned off do not track ray
interactions). After the user selects the active windows, they proceed
to shut off and their interactions get disabled.

Still have to figure out how to set it up for a given number of trials.
Time measurement is next on the list.